Add  Q and A plugins
Based on https://github.com/ColdTrick/questions release #v2.1.2
The only difference is the addition of the French language and fix
minor typo in English. Which there is the same as the content in the
repo
https://github.com/gctools-hackathon-ircc/questions/tree/elgg_112_gcconn
ex

The PR https://github.com/gctools-outilsgc/gcconnex/pull/952 need to be
merged before this one has it fix a displaying issue.

FYI - [French language
support](https://github.com/ColdTrick/questions/pull/11) has been
submitted upstream with the [fix for English
typo](https://github.com/ColdTrick/questions/pull/12) but for that may
only be applied for version 5+ of the
https://github.com/ColdTrick/questions plugin which require a more
recent version of elgg