Mobile API: Get_Groups - Description, Access, Clean_Text
As used in get_group, get_groups now checks if group is public or if user is a member. If no to both, will set description to access:private message.
Clean text for group lists. Same as usergroups.
Truncated the description, as full description is not needed for group lists.