Get_Newsfeed: Comments (#1875)
* Get_Newsfeed2: Container Variables

For comments, $other is used to pass the comment's container's type and guid. Allows for mobile application to have context for the comment location.

* Get_Newsfeed: Remove Unneeded If-Else

If was checking to make sure User_Entity exists, but this is handled above. Begining of function already checks for existing, and that it is a ElggUser.
Else is uneeded because it would never be reached.