Mobile get_events: Now to Future, Never Past
`$now = time()` variable moved to sooner to be used in user input 'from' handling.
If there is a from variable passed in, also checks that it is greater than now. If neither of those, will default to using now. Therefore it will never check for events that are in the past.